Prevent error from #198
This only prevents the error from #198, but still doesn't actually detect the loading area because the tension belts are in a vertical plane.  There is also something really weird going on with a delay when switching to this platform for the first time.  I've added some timestamps to help debug, but it seems to happen between `onLoad` and `doUpdate`.